Issue-965: Custom YAML Mapper fix.
